reorganizes component folder structure
