+++ src/component/App/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Routes, Route, Link } from 'react-router-dom';
 import './App.css';
 import homeImg from './assets/images/home.png';
 import newImg from './assets/images/new.png';
-import Home from './component/Home';
+import Home from '../Home/Home';
 
 function App() {
 	return (
